chore(migration): dashboard-fonts.spec to ts
