--- tests/dashboard/dashboard-fonts.spec.js
@@ -1,75 +1,0 @@
-const { mainTest } = require('../../fixtures');
-const { DashboardPage } = require('../../pages/dashboard/dashboard-page');
-const { random } = require('../../helpers/string-generator');
-const { TeamPage } = require('../../pages/dashboard/team-page');
-const { qase } = require('playwright-qase-reporter/playwright');
-const { createTeamName } = require('../../helpers/teams/create-team-name');
-
-const teamName = createTeamName();
-
-let teamPage, dashboardPage;
-
-mainTest.beforeEach(async ({ page }) => {
-  teamPage = new TeamPage(page);
-  dashboardPage = new DashboardPage(page);
-  await teamPage.createTeam(teamName);
-  await teamPage.isTeamSelected(teamName);
-});
-
-mainTest(qase(1152, 'Upload single font'), async () => {
-  await dashboardPage.openSidebarItem('Fonts');
-  await dashboardPage.uploadFont('fonts/Pacifico.ttf');
-  await dashboardPage.isFontExists('Pacifico', 'Regular');
-});
-
-mainTest(qase(1153, 'Fonts - upload (multiple)'), async () => {
-  await dashboardPage.openSidebarItem('Fonts');
-  await dashboardPage.uploadAllFonts([
-    'fonts/Pacifico.ttf',
-    'fonts/SourceCodePro-Regular.woff',
-  ]);
-  await dashboardPage.areFontsListed([
-    { fontName: 'Pacifico', fontStyle: 'Regular' },
-    { fontName: 'Source Code Pro', fontStyle: 'Regular' },
-  ]);
-});
-
-mainTest(qase(1154, 'Fonts - upload fail invalid file format'), async () => {
-  await dashboardPage.openSidebarItem('Fonts');
-  await dashboardPage.uploadFontWithInvalidFormat('images/images.png');
-});
-
-mainTest(qase(1155, 'Search font'), async () => {
-  await dashboardPage.openSidebarItem('Fonts');
-  await dashboardPage.uploadFont('fonts/ArialTh.ttf');
-  await dashboardPage.uploadFont('fonts/Allura-Regular.otf');
-  await dashboardPage.searchFont('Arial Th');
-  await dashboardPage.isFontExists('Arial Th', 'Regular');
-  await dashboardPage.isFontNotExist('Allura-Regular');
-});
-
-mainTest(qase(1157, 'Edit font BUG'), async () => {
-  await dashboardPage.openSidebarItem('Fonts');
-  await dashboardPage.uploadFont('fonts/Allura-Regular.otf');
-  await dashboardPage.isFontExists('Allura', 'Regular');
-  await dashboardPage.editFont('New Test Font');
-  await dashboardPage.isFontExists('New Test Font', 'Regular');
-});
-
-mainTest(qase(1158, 'Delete font'), async () => {
-  await dashboardPage.openSidebarItem('Fonts');
-  await dashboardPage.uploadFont('fonts/Pacifico.ttf');
-  await dashboardPage.isFontExists('Pacifico', 'Regular');
-  await dashboardPage.deleteFont();
-  await dashboardPage.isFontsTablePlaceholderDisplayed(
-    'Custom fonts you upload will appear here.',
-  );
-});
-
-mainTest(qase(1159, 'Delete font - Cancel button check'), async () => {
-  await dashboardPage.openSidebarItem('Fonts');
-  await dashboardPage.uploadFont('fonts/Pacifico.ttf');
-  await dashboardPage.isFontExists('Pacifico', 'Regular');
-  await dashboardPage.cancelDeleteFont();
-  await dashboardPage.isFontExists('Pacifico', 'Regular');
-});
